--- a/microbenchmarks/large_import_bench.py
+++ b/microbenchmarks/large_import_bench.py
@@ -1,0 +1,16 @@
+import re
+import os
+import sre_compile
+import weakref
+import threading
+import abc
+import _abcoll
+import collections
+import multiprocessing
+import argparse
+import bisect
+import heapq
+import inspect
+import netrc
+import pickle
+import socket