Some small optimizations speed up this part of the pipeline ~25%.  Not a huge overall win since refcounting is about 20% of (cached-case) LLVM time, but at least we can claw some of the refcounting regression back.

The remaining time is largely spent manipulating the LLVM IR.

```
           django_template3.py             2.3s (2)             2.3s (2)  -0.6%
                 pyxl_bench.py             2.2s (2)             2.2s (2)  -0.4%
     sqlalchemy_imperative2.py             2.5s (2)             2.5s (2)  -0.9%
                pyxl_bench2.py             1.2s (2)             1.2s (2)  -0.8%
             pyxl_bench_10x.py            16.8s (2)            16.8s (2)  -0.4%
       django_template3_10x.py            12.4s (6)            12.3s (2)  -0.4%
 sqlalchemy_imperative2_10x.py            18.0s (2)            18.1s (2)  +0.1%
            pyxl_bench2_10x.py            10.3s (2)            10.2s (2)  -1.0%
                       geomean                 5.3s                 5.3s  -0.5%
```